FIX: Avoid optional sklearn import during test collection
