+++ mne/decoding/tests/test_search_light.py
@@ -22,7 +22,7 @@ from sklearn.pipeline import make_pipeline
 from sklearn.svm import SVC
 
 from mne.decoding.search_light import GeneralizingEstimator, SlidingEstimator
-from mne.decoding.tests import parametrize_with_checks
+from mne.decoding.tests._sklearn import parametrize_with_checks
 from mne.decoding.transformer import Vectorizer
 from mne.utils import check_version, use_log_level
