+++ mne/decoding/tests/test_xdawn.py
@@ -9,7 +9,7 @@ from numpy.testing import assert_array_almost_equal
 pytest.importorskip("sklearn")
 
 from mne.decoding import XdawnTransformer, read_xdawn_transformer
-from mne.decoding.tests import parametrize_with_checks
+from mne.decoding.tests._sklearn import parametrize_with_checks
 
 
 @pytest.mark.filterwarnings("ignore:.*Only one sample available.*")
